Move Marshal to shared partition
